issue: 3514044 Moving cq_mgr_regrq RX common to cq_mgr
