+++ src/core/dev/cq_mgr_strq.cpp
@@ -60,7 +60,7 @@ cq_mgr_strq::cq_mgr_strq(ring_simple *p_ring, ib_ctx_handler *p_ib_ctx_handler,
                          uint32_t strides_num,
                          struct ibv_comp_channel *p_comp_event_channel,
                          bool call_configure)
-    : cq_mgr_regrq(p_ring, p_ib_ctx_handler, cq_size, p_comp_event_channel, true, call_configure)
+    : cq_mgr(p_ring, p_ib_ctx_handler, cq_size, p_comp_event_channel, true, call_configure)
     , _owner_ring(p_ring)
     , _stride_size_bytes(stride_size_bytes)
     , _strides_num(strides_num)
